term_project/requirements.txt
keras
opencv-python
tensorflow-gpu
Pillow
matplotlib